requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
psutil>=5.8.0
tqdm>=4.0.0 